// src/app/pages/login/user-service/model/user.ts
export interface User {
  username: string;
  userId: number;
  firstName: string;
  lastName: string;
  points: number;
  avatar: string;
  desc: string;
  games: number;
  email: string;
  HP_learningFactor: number;
  HS_learningFactor: number;
  NC_learningFactor: number;
  PA_learningFactor: number;
  SC_learningFactor: number;
  SE_learningFactor: number;
}