// src/app/pages/quiz/models/learning-mode-api.ts
export interface LearningModeApi {

  poprawneOdpowiedzi_HP: number;
  poprawneOdpowiedzi_HS: number;
  poprawneOdpowiedzi_NC: number;
  poprawneOdpowiedzi_PA: number;
  poprawneOdpowiedzi_SC: number;
  poprawneOdpowiedzi_SE: number;

  niepoprawneOdpowiedzi_HP: number;
  niepoprawneOdpowiedzi_HS: number;
  niepoprawneOdpowiedzi_NC: number;
  niepoprawneOdpowiedzi_PA: number;
  niepoprawneOdpowiedzi_SC: number;
  niepoprawneOdpowiedzi_SE: number;
}